Refactor view class structure
