+++ src/view/customframes/SudokugridFrame.py
@@ -41,7 +41,7 @@ class SudokuFrame(ctk.CTkFrame):
 
     def initialize_fields(self):
     
-        self.game_entry_dimension = int(self.master.master.window_height // self.game_gridsize * 0.75)
+        self.game_entry_dimension = int(self.window_size // self.game_gridsize * 0.75)
 
         self.sudoku_row = 0
         for row in range(self.gridsize):
